Move the parts for using Word documents as Jinja templates into JinjaService
